Move utils into api package to fix cyclic deps
