--- internal/utils/string.go
@@ -1,7 +1,0 @@
-package utils
-
-import "strings"
-
-func TrimString(s string) string {
-	return strings.Trim(s, " \t\n\r")
-}
